Made capability props configurable
